Split the Dialect 3 reader into many smaller files.
